Develop/main

+++ docker/openapi/swagger-merger/generateJsonSchema.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 
+# 型定義ファイル格納先
 readonly TARGET_PATH=$1
+# JSON Schema作成先
+readonly DESTINATION_PATH=$2
+# 除外条件(ファイル名を正規表現で)
+readonly EXCLUSION_CONDITION=$3
 readonly TMP_FILE=$(dirname $0)/.tmp
 readonly INTERVAL=1
 
@@ -8,6 +13,12 @@ if [ -z $TARGET_PATH ] || [ ! -d $TARGET_PATH ]; then
   echo "[ERROR]Please specify target path as FIRST argument."
   exit
 fi
+
+if [ -z $DESTINATION_PATH ] || [ ! -d $DESTINATION_PATH ]; then
+  echo "[ERROR]Please specify destination path as SECOND argument."
+  exit
+fi
+
 
 trap "rm -f $TMP_FILE" 1 2 3 15
 
@@ -23,6 +34,12 @@ done
 
 while true; do
   for f in $(find $TARGET_PATH -type f -name "*.ts"); do
+    ## 除外条件を指定している & 条件に合致したら処理を飛ばす
+    ! -z $EXCLUSION_CONDITION && \
+      echo $f | grep -qE $EXCLUSION_CONDITION && \
+      continue
+
+
     current=$(ls --full-time $f | awk '{print $6$7}')
     last=$(grep -e "${f}@@" $TMP_FILE | awk -F"@@" '{print $2}')
 
@@ -41,7 +58,7 @@ while true; do
         echo "${f}@@${current}" >> $TMP_FILE
       fi
 
-      typescript-json-schema $f $(basename $f | cut -d. -f1) --required > $(echo $f | sed -e s/\.ts$/.json/)
+      typescript-json-schema $f $(basename $f | cut -d. -f1) --required > ${DESTINATION_PATH}/$(basename $f | sed -e s/\.ts$/.json/)
     fi
   done
